requirements.txt
babel==2.16.0
GPUtil==1.4.0
numpy==1.26.4
pillow==10.4.0
psutil==5.9.8
pyadl==0.1
pyserial==3.5
pythonnet==3.0.4
pywin32==306
PyYAML==6.0.2
spotipy==2.24.0
uptime==3.0.1
flask == 3.0.3